applications/worker-service/requirements.txt
google-cloud-pubsub==2.19.0
google-cloud-storage==2.14.0
google-cloud-bigquery==3.17.1
google-cloud-logging==3.9.0
google-cloud-secret-manager==2.17.0
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.0.1
